feat: add custom dotprop version to fix ESM issue
